Using py::object to pass tensors to the matmul method, allowing for dimensionality checks
